Separate MoEGate op (#44203)
Cleanup

The optimized MoEGate op relied on `generic_op`. This could hinder the
generalization ability.
This PR makes the op a separate op without relying on `generic_op`
